Allow CSI driver to function entirely within container
We have been hitting numerous customer issues due to the state
that is left behind on a node when we install the CSI driver
using the current process, especially due to apt having issues
with differing versions of the driver.

This change removes the need for the driver to be installed
directly into the node's filesystem. Since the lustre driver
is installed into the kernel as a module, it is sufficient
for our driver container to have the appropriate permissions
and mount points solely for loading the module and operating
as a CSI driver.

This driver is configured to unload the kernel module when
the pod is deleted, allowing for a seamless upgrade experience
in future versions, where the kernel module is loaded from
a clean system, and without leaving permanent installation
artifacts around in the node's root filesystem.