Add sub-dir functionality
This commit adds the ability for pods to bind to a Lustre volume
in a specified subdirectory. The pod that binds the volume in
this way can only see that path and the paths under it. It has
no access or visibility to other paths above it in the Lustre
filesystem.

This has been implemented similarly to other CSI drivers to meet
user expectations, but has the added benefit of allowing the
user to leverage metadata from both the pvc/pv metadata and
the pod's metadata as well.

Numerous checks have been put in place to ensure that data
outside of the requested subdirectory is not deleted and that
read only volumes are handled reasonably.

The long-haul tests have had compatible subdirectory values
set to exercise this functionality while still maintaining
its expectations as to data persistence, etc.