Fix flaky test
The parameters in the request do not have a guaranteed order. So,
in one test, the regex assertion would fail if the invalid
parameters were not listed in the original order. Since we do not
care what order the error reports the parameters in, this commit
changes the assertion only to ensure the expected unique values are
present in the error.