Simplify logic when ensuring strict subpath
By trimming the sub-dir values to always be relative before use,
we can leverage the existing filepath.IsLocal predicate to simplify
the logic we have deal with when determining if a sub-dir value
will attempt to break out of the base directory. Since we still
want to ensure that the subdirectory is not exqual to the base,
we still explicitly make that check, so the helper method to
abstract away that logic remains.