Revert metadata change until release image is available
This change went in with the sub-dir changes, but it
requires the new image for the driver to function
properly. We need to change this value back until we
can update the deployment files to use the new driver
image. As it is, the dynamic volume creation fails
because the extra metadata is being passed to a
controller that is not yet expecting it.

Commenting out sub-dir references in tests for
similar reasons.