Remove retain-sub-dir option
The retain-sub-dir feature is intended to allow the user to have
their subdirectory data automatically cleaned up when a pod
unmounts from the Lustre filesystem. However, based on test,
in its current form it allows the user to easily get themselves
into a bad state or to lose data. For this release, we will
maintain the subdirectory creation and mounting logic, but
will wait on any customer feedback before re-enabling this
functionality based on expected use cases.