Use golangci-lint config
Instead of having different linters and configuration options in
the various places we call golangci-lint, this changes how we
run it so that it uses a shared configuration. This prevents the
case where one of these checks would pass only to fail for the user
when the workflows are run prior to merging.