Check for all Error, Fail, or Timeout codes in 200 responses
We were seeing errors of this form in 200 responses:
{
  "startTime": "2025-07-29T20:54:50.1830248+00:00",
  "endTime": "2025-07-29T20:55:05.0436511+00:00",
  "status": "Failed",
  "error": {
    "code": "AscInternalError",
    "message": "Provisioning timed out deploying Cluster VMs"
  },
  "name": "994d67b9-590b-47c4-9bd4-bc83771c1f6d"
}

Since we were only checking for a couple of specific errors,
the driver did not know that this was a failure case and
should be deleted and retried.

This commit adds logic to check the cluster state for delete
and retry if any of 'Error', 'Fail', or 'Timeout' appear
in the error code.